fix: resolve clippy warnings from rebase
- Add missing SMBError import in ipc.rs
- Move doc comments inside bitflags! macros
- Remove unused SMBByteSize import
- Fix unnecessary cast and needless borrow
- Allow too_many_arguments on FileAllInformation::new (matches MS-FSCC spec)

+++ smb/src/server/open.rs
@@ -141,7 +141,7 @@ impl<S: Server> Open for SMBOpen<S> {
     }
 
     fn file_id(&self) -> SMBFileId {
-        SMBFileId::new(self.global_id as u64, self.session_id as u64)
+        SMBFileId::new(self.global_id as u64, self.session_id)
     }
 
     fn file_metadata(&self) -> SMBResult<SMBFileMetadata> {
